move cmp impls into submodule
